refactor(pool-page): extract _components/pool-header
Move the PoolHeader component (~135 lines) out of page.tsx into
_components/pool-header.tsx. PoolHeader takes only props and is
self-contained — clean cut.

Behavior-preserving — verified via:
- pnpm test (1366/1366 green)
- pnpm typecheck (clean)
- chrome-devtools snapshot of /pool/<id>?tab=reserves vs baseline:
  zero structural changes, console clean

page.tsx: 2348 → 2211 lines (-137).
